+++ packages/actions/src/contracts/epoch-manager.ts
@@ -1,0 +1,15 @@
+import { epochManagerABI } from '@celo/abis-12'
+import { getContract, GetContractReturnType, PublicClient, WalletClient } from 'viem'
+import { resolveAddress } from './registry'
+
+export async function getEpochManagerContract<T extends WalletClient | PublicClient = PublicClient>(
+  client: T
+): Promise<EpochManager<T>> {
+  return getContract({
+    address: await resolveAddress(client, 'EpochManager'),
+    abi: epochManagerABI,
+    client,
+  })
+}
+export type EpochManager<T extends WalletClient | PublicClient = PublicClient> =
+  GetContractReturnType<typeof epochManagerABI, T>
