+++ packages/actions/src/index.ts
@@ -1,0 +1,3 @@
+export { resolveAddress } from './contracts/registry.js'
+export { getGasPriceOnCelo } from './rpc-methods.js'
+export { ContractName } from './types/contract-name.js'
