+++ packages/actions/src/multicontract-interactions/stake/index.ts
@@ -1,0 +1,1 @@
+export { ElectedRpcNode, getElectedRpcNodes } from './elected-rpc-nodes'
