+++ packages/actions/src/setup/index.ts
@@ -1,0 +1,2 @@
+export * from './chains.js'
+export * from './client.js'
