+++ packages/cli/package.json
@@ -39,6 +39,7 @@
     "cs": "yarn --top-level run cs"
   },
   "dependencies": {
+    "@celo/actions": "0.0.0",
     "@celo/abis": "11.0.0",
     "@celo/base": "^7.0.2",
     "@celo/compliance": "1.0.28",
@@ -77,7 +78,7 @@
     "humanize-duration": "^3.32.1",
     "path": "^0.12.7",
     "prompts": "^2.0.1",
-    "viem": "^2.23.5",
+    "viem": "^2.29.2",
     "web3": "1.10.4"
   },
   "devDependencies": {
