+++ packages/cli/src/commands/transfer/reals.ts
@@ -1,4 +1,4 @@
-import { getStableTokenBrlContract } from '../../packages-to-be/contracts'
+import { getStableTokenBrlContract } from '../../packages-to-be/stable-tokens'
 import { TransferStableBase } from '../../transfer-stable-base'
 
 export default class TransferEuros extends TransferStableBase {
