+++ packages/cli/src/packages-to-be/getEpochInfo.ts
@@ -1,5 +1,5 @@
+import { getEpochManagerContract } from '@celo/actions/epoch-manager-contract'
 import { PublicClient } from 'viem'
-import { getEpochManagerContract } from './contracts'
 
 export async function getEpochInfo(client: PublicClient) {
   const epochManager = await getEpochManagerContract(client)
