+++ packages/cli/src/packages-to-be/validators.ts
@@ -1,10 +1,10 @@
 import { accountsABI, lockedGoldABI, validatorsABI } from '@celo/abis-12'
+import { resolveAddress } from '@celo/actions'
+import { getValidatorsContract } from '@celo/actions/validators-contract'
 import { Address, concurrentMap, ensureLeading0x, eqAddress, StrongAddress } from '@celo/base'
 import { fromFixed } from '@celo/utils/lib/fixidity'
 import BigNumber from 'bignumber.js'
 import { PublicClient } from 'viem'
-import { resolveAddress } from './address-resolver'
-import { getValidatorsContract } from './contracts'
 import { bigintToBigNumber } from './utils'
 
 export interface Validator {
