+++ packages/cli/src/transfer-stable-base.ts
@@ -1,15 +1,15 @@
+import { getGasPriceOnCelo } from '@celo/actions'
+import { getCeloERC20Contract } from '@celo/actions/celo-contract'
+import { getERC20Contract } from '@celo/actions/erc20-contract'
 import { Flags } from '@oclif/core'
 import { isAddressEqual, PublicClient } from 'viem'
 import { BaseCommand } from './base'
 import {
-  getERC20Contract,
-  getGoldTokenContract,
   StableToken,
   StableTokenContract,
   StableTokenContractGetter,
   StableTokens,
-} from './packages-to-be/contracts'
-import { getGasPriceOnCelo } from './packages-to-be/utils'
+} from './packages-to-be/stable-tokens'
 import { newCheckBuilder } from './utils/checks'
 import { displaySendViemContractCall, failWith } from './utils/cli'
 import { CustomFlags } from './utils/command'
@@ -66,7 +66,7 @@ export abstract class TransferStableBase extends BaseCommand {
     } catch {
       failWith(`The ${stableToken} token was not deployed yet`)
     }
-    const goldTokenContract = await getGoldTokenContract(client as PublicClient)
+    const goldTokenContract = await getCeloERC20Contract(client as PublicClient)
 
     const stableContractData = {
       abi: stableTokenContract.abi,
