+++ packages/cli/src/utils/cli.ts
@@ -16,8 +16,8 @@ import { formatEther } from 'ethers/lib/utils'
 import humanizeDuration from 'humanize-duration'
 import { convertEthersToCeloTx } from './mento-broker-adaptor'
 
+import { CeloClient, WalletCeloClient } from '@celo/actions/setup'
 import { CeloTransactionRequest } from 'viem/celo'
-import { CeloClient, WalletCeloClient } from '../packages-to-be/client'
 
 const CLIError = Errors.CLIError
